# eduverse_server/models/error.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from eduverse_server.models.base_model import Model
from eduverse_server import util


class Error(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, error=None, code=None, message=None):  # noqa: E501
        """Error - a model defined in OpenAPI

        :param error: The error of this Error.  # noqa: E501
        :type error: bool
        :param code: The code of this Error.  # noqa: E501
        :type code: int
        :param message: The message of this Error.  # noqa: E501
        :type message: str
        """
        self.openapi_types = {
            'error': bool,
            'code': int,
            'message': str
        }

        self.attribute_map = {
            'error': 'error',
            'code': 'code',
            'message': 'message'
        }

        self._error = error
        self._code = code
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'Error':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Error of this Error.  # noqa: E501
        :rtype: Error
        """
        return util.deserialize_model(dikt, cls)

    @property
    def error(self) -> bool:
        """Gets the error of this Error.


        :return: The error of this Error.
        :rtype: bool
        """
        return self._error

    @error.setter
    def error(self, error: bool):
        """Sets the error of this Error.


        :param error: The error of this Error.
        :type error: bool
        """

        self._error = error

    @property
    def code(self) -> int:
        """Gets the code of this Error.


        :return: The code of this Error.
        :rtype: int
        """
        return self._code

    @code.setter
    def code(self, code: int):
        """Sets the code of this Error.


        :param code: The code of this Error.
        :type code: int
        """

        self._code = code

    @property
    def message(self) -> str:
        """Gets the message of this Error.


        :return: The message of this Error.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this Error.


        :param message: The message of this Error.
        :type message: str
        """

        self._message = message