# eduverse_server/models/grade.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from eduverse_server.models.base_model import Model
from eduverse_server import util


class Grade(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, subject_id=None, grade=None):  # noqa: E501
        """Grade - a model defined in OpenAPI

        :param subject_id: The subject_id of this Grade.  # noqa: E501
        :type subject_id: str
        :param grade: The grade of this Grade.  # noqa: E501
        :type grade: int
        """
        self.openapi_types = {
            'subject_id': str,
            'grade': int
        }

        self.attribute_map = {
            'subject_id': 'subject_id',
            'grade': 'grade'
        }

        self._subject_id = subject_id
        self._grade = grade

    @classmethod
    def from_dict(cls, dikt) -> 'Grade':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Grade of this Grade.  # noqa: E501
        :rtype: Grade
        """
        return util.deserialize_model(dikt, cls)

    @property
    def subject_id(self) -> str:
        """Gets the subject_id of this Grade.


        :return: The subject_id of this Grade.
        :rtype: str
        """
        return self._subject_id

    @subject_id.setter
    def subject_id(self, subject_id: str):
        """Sets the subject_id of this Grade.


        :param subject_id: The subject_id of this Grade.
        :type subject_id: str
        """

        self._subject_id = subject_id

    @property
    def grade(self) -> int:
        """Gets the grade of this Grade.


        :return: The grade of this Grade.
        :rtype: int
        """
        return self._grade

    @grade.setter
    def grade(self, grade: int):
        """Sets the grade of this Grade.


        :param grade: The grade of this Grade.
        :type grade: int
        """

        self._grade = grade