# eduverse_server/models/homework.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from eduverse_server.models.base_model import Model
from eduverse_server import util


class Homework(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, subject_id=None, title=None, description=None, deadline=None):  # noqa: E501
        """Homework - a model defined in OpenAPI

        :param subject_id: The subject_id of this Homework.  # noqa: E501
        :type subject_id: str
        :param title: The title of this Homework.  # noqa: E501
        :type title: str
        :param description: The description of this Homework.  # noqa: E501
        :type description: str
        :param deadline: The deadline of this Homework.  # noqa: E501
        :type deadline: date
        """
        self.openapi_types = {
            'subject_id': str,
            'title': str,
            'description': str,
            'deadline': date
        }

        self.attribute_map = {
            'subject_id': 'subject_id',
            'title': 'title',
            'description': 'description',
            'deadline': 'deadline'
        }

        self._subject_id = subject_id
        self._title = title
        self._description = description
        self._deadline = deadline

    @classmethod
    def from_dict(cls, dikt) -> 'Homework':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Homework of this Homework.  # noqa: E501
        :rtype: Homework
        """
        return util.deserialize_model(dikt, cls)

    @property
    def subject_id(self) -> str:
        """Gets the subject_id of this Homework.


        :return: The subject_id of this Homework.
        :rtype: str
        """
        return self._subject_id

    @subject_id.setter
    def subject_id(self, subject_id: str):
        """Sets the subject_id of this Homework.


        :param subject_id: The subject_id of this Homework.
        :type subject_id: str
        """

        self._subject_id = subject_id

    @property
    def title(self) -> str:
        """Gets the title of this Homework.


        :return: The title of this Homework.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Homework.


        :param title: The title of this Homework.
        :type title: str
        """

        self._title = title

    @property
    def description(self) -> str:
        """Gets the description of this Homework.


        :return: The description of this Homework.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Homework.


        :param description: The description of this Homework.
        :type description: str
        """

        self._description = description

    @property
    def deadline(self) -> date:
        """Gets the deadline of this Homework.


        :return: The deadline of this Homework.
        :rtype: date
        """
        return self._deadline

    @deadline.setter
    def deadline(self, deadline: date):
        """Sets the deadline of this Homework.


        :param deadline: The deadline of this Homework.
        :type deadline: date
        """

        self._deadline = deadline