# eduverse_server/models/subject.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from eduverse_server.models.base_model import Model
from eduverse_server import util


class Subject(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, subject_id=None, name=None):  # noqa: E501
        """Subject - a model defined in OpenAPI

        :param subject_id: The subject_id of this Subject.  # noqa: E501
        :type subject_id: str
        :param name: The name of this Subject.  # noqa: E501
        :type name: str
        """
        self.openapi_types = {
            'subject_id': str,
            'name': str
        }

        self.attribute_map = {
            'subject_id': 'subject_id',
            'name': 'name'
        }

        self._subject_id = subject_id
        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'Subject':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Subject of this Subject.  # noqa: E501
        :rtype: Subject
        """
        return util.deserialize_model(dikt, cls)

    @property
    def subject_id(self) -> str:
        """Gets the subject_id of this Subject.


        :return: The subject_id of this Subject.
        :rtype: str
        """
        return self._subject_id

    @subject_id.setter
    def subject_id(self, subject_id: str):
        """Sets the subject_id of this Subject.


        :param subject_id: The subject_id of this Subject.
        :type subject_id: str
        """

        self._subject_id = subject_id

    @property
    def name(self) -> str:
        """Gets the name of this Subject.


        :return: The name of this Subject.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Subject.


        :param name: The name of this Subject.
        :type name: str
        """

        self._name = name