# eduverse_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API платформы EduVerse. Авторы: команда 5
  title: EduVerse
  version: 1.0.0
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/comand_5/API_EduVerse/1.0.0
security:
- bearerAuth: []
...
paths:
  /academic/grades:
    get:
      description: Получение списка оценок конкретного студента
      operationId: get_list_grades
      parameters:
      - description: Идентификатор студента
        explode: true
        in: query
        name: student_id
        required: true
        schema:
          type: string
        style: form
      - description: Номер страницы
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Количество элементов на странице
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Grade"
                type: array
          description: Список оценок
        "404":
          description: Оценки не найдены
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Оценки студента
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
    post:
      description: Доступно только преподавателю
      operationId: create_grade
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Grade"
        required: true
      responses:
        "200":
          description: Оценка сохранена
        "404":
          description: Оценки не найдены
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Добавить оценку
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
  /academic/homework:
    get:
      operationId: get_list_homework
      parameters:
      - explode: true
        in: query
        name: subject_id
        required: true
        schema:
          type: string
        style: form
      - description: Номер страницы
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Количество элементов на странице
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Homework"
                type: array
          description: Список заданий
        "404":
          description: Домашняя работа не найдена
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Список домашнего задания
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
    post:
      description: Доступно преподавателю
      operationId: create_homework
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Homework"
        required: true
      responses:
        "201":
          description: Задание добавлено
        "404":
          description: домашнее задание не найдено
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Добавить домашнее задание
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
  /academic/subjects:
    get:
      operationId: get_list_subject
      parameters:
      - description: Номер страницы
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Количество элементов на странице
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Subject"
                type: array
          description: Список предметов
        "404":
          description: Предметы не найдены
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Список предметов
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
    post:
      operationId: create_subject
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subject"
        required: true
      responses:
        "201":
          description: Предмет создан
        "404":
          description: Предметы не найдены
        "401":
          description: Неавторизованный доступ
        "500":
          description: Ошибка сервера
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Произошла ошибка
      summary: Добавить предмет
      tags:
      - Academic
      x-openapi-router-controller: eduverse_server.controllers.academic_controller
...
components:
...
  schemas:
    Error:
      example:
        code: 404
        error: true
        message: Ресурс не найден
      properties:
        error:
          example: true
          title: error
          type: boolean
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Ресурс не найден
          title: message
          type: string
      title: Error
      type: object
...
    Homework:
      example:
        subject_id: subject_id
        description: description
        title: title
        deadline: 2000-01-23
      properties:
        subject_id:
          title: subject_id
          type: string
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        deadline:
          format: date
          title: deadline
          type: string
      title: Homework
      type: object
...
    Subject:
      example:
        subject_id: sub_1
        name: Математика
      properties:
        subject_id:
          example: sub_1
          title: subject_id
          type: string
        name:
          example: Математика
          title: name
          type: string
      title: Subject
      type: object
    Grade:
      example:
        subject_id: sub_1
        grade: 5
      properties:
        subject_id:
          example: sub_1
          title: subject_id
          type: string
        grade:
          example: 5
          title: grade
          type: integer
      title: Grade
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: eduverse_server.controllers.security_controller.info_from_bearerAuth